Finish unit tests for Player class.
The Player class is now fully tested, and the methods come out a little
cleaner I think. However, the integration tests no longer work. This is
is to do with the `forget` method. By the changes in this commit, the
hospital-resident algorithms fall over. Will require unit testing on the
algorithm components.