Dump commit. Change player match to matching.
Trying to sort out the Solver matching setter method. Want any change to
a matching to be done through a specific method that is tested, and for
that method to also update the suitor/reviewer objects in question.

Likely solution is that the algorithms and current contents of util.py
should be methods in the specific and general Solver classes
respectively.