Refactor to classes (#29)
* Draft new scripts. Must update test suite in morn.

* Finish docstrings.

* Blackened source. Moving to tests.

* Write tests for hospital-resident algorithms.

* Clean up HR tests.

* Write tests for SM games.

* Start writing Player unit tests.

* Finish unit tests for Player class.

The Player class is now fully tested, and the methods come out a little
cleaner I think. However, the integration tests no longer work. This is
is to do with the `forget` method. By the changes in this commit, the
hospital-resident algorithms fall over. Will require unit testing on the
algorithm components.

* Add additional (unnecessary?) return None statemts

* Change elif to if statement in Player method.

* Get integration tests to pass. Not the best soln.

* Update player forget method docstring.

* Update README.rst

* Make matching and unmatching Player methods.

* Flatten integration test suite structure.

* Dump commit. Stopping work for the day.

* Restructure of code base and test suite.

* Dump commit. Change player match to matching.

Trying to sort out the Solver matching setter method. Want any change to
a matching to be done through a specific method that is tested, and for
that method to also update the suitor/reviewer objects in question.

Likely solution is that the algorithms and current contents of util.py
should be methods in the specific and general Solver classes
respectively.

* Grab last of player attribute change.

* Update unit tests for base classes.

* Rtc solver tests (#28)

* Start working toward 100% coverage in unit tests.

* Add Matcher class and prepare exisiting tests.

* Implement Matching class. Full coverage.

* Start implement of HospRes solver.

* Get to 100% coverage on SM solver. Last test bad.

* Finalise stable marriage solver.

* Format source code with Black.

* Start HR unit tests.

* Continue with HR tests. Done for day.

* Correct follow-on error in HR solver test.

* Update solver docstrings for Matching object.

* Finish HR solver tests.