Incorporate player copying (#83)
* Alter tests to support copies of players.

Player instances are now internal to a Game instance so the tests
have to reflect that. Without designing some overly elaborate __eq__ and
__hash__ methods, this seems sensible for now (other aspects of a Game
instance require the Player class to be hashable).

* Implement player copying.

Player instances are copied at the instantiation of a Game subclass.
This is done using `copy.deepcopy`. That way, multiple Game instances
can be created from one set of player lists.