JOSS review (#97)
* Incorporate player copying (#83)

* Alter tests to support copies of players.

Player instances are now internal to a Game instance so the tests
have to reflect that. Without designing some overly elaborate __eq__ and
__hash__ methods, this seems sensible for now (other aspects of a Game
instance require the Player class to be hashable).

* Implement player copying.

Player instances are copied at the instantiation of a Game subclass.
This is done using `copy.deepcopy`. That way, multiple Game instances
can be created from one set of player lists.

* Add hypothesis to setup.py (#87)

* Add test for inconsistent coverage. (#86)

Occasionally, coverage falls below 100% because of an edge case in HR
where no resident is removed from consideration for forgetting all of
their preferences. This simple example test rectifies that.

* Add hypothesis to `tests_require` in `setup.py`

* Make Game a metaclass (#94)

* Reimplement `test_game.py` for `abc.ABCMeta`.

* Change `Game` to `BaseGame` as `ABCMeta` subclass.

* Change inheritance in each game.

* Format with black, blackbook and isort.

* Format with black, blackbook and isort.

* Edit paper for JOSS review (#93)

* Add test for inconsistent coverage. (#86)

Occasionally, coverage falls below 100% because of an edge case in HR
where no resident is removed from consideration for forgetting all of
their preferences. This simple example test rectifies that.

* Fix flaky `Player.forget()` behaviour (#88)

* Add `ValueError` to exceptions in SR(2) test.

Since the first phase is not being completed, the preferences have not
been reduced and the all-or-nothing cycles can contain repeats of pairs.
This means players attempt to forget someone they've already forgotten.

* Remove the try-except from `Player.forget()`.

No longer necessary as the tests catch this edge case (that would not
ever actually happen in SR).

* Flesh out example paragraph.

* Remove fluff around edu infrastructure.

* Fix documentation for JOSS review (#85)

* Add sphinxcontrib-bibtex to docs/env file.

* Remove "Search Page" from "Indices and Tables"

* Add notebook pointer prologue.

* Clarify installation instructions for pyversion.

* Move install intro to a note directive.

* Update README (and supporting files) (#84)

* Add CONTRIBUTING file to the top level.

* Remove duplicate "Documentation" sect from README

* Change "PR" to "pull request" in README

* Add python_requires parameter to setup (#92)