Developing v1.3.3 (#127)
* Fix ipython version to avoid warnings. (#126)

* Fix ipython version to avoid warnings.

There were some weird warnings coming from an internal
`ipython`/`ipykernel` API, causing `pytest-nbval` to fail its tests. So,
downgrading `ipython` to 7.10 until that is fixed.

Details are available under
[this issue](https://github.com/ipython/ipykernel/issues/540).

* Fix ipython version in workflow file.

* Move unmatching inside `second_phase` of SR (#125)

* Move unmatching inside `second_phase`

This was causing simple examples to fail because the second phase of the
algorithm never kicked in. By moving the unmatching, this is avoided,
and is fine because of the corollaries in Irving's paper.

* Add example tests from #124

* Checkout CI file from dev.